Introduce model instance and new temporal structure
Re #120

The model start, end, and roll forward are now read from the model
instance. This is used to generate rolling windows, and then
time slices for each of these windows within a simple for loop.

The database file DataStoreA4.sqlite has been added to
examples/data. I know this goes against the rule of not adding data
to git, but I think it is very helpful to have a test case which
we can test our changes against. Of course, DataStoreA4.sqlite
uses the new temporal structure with the model instance. The model
can be run by including test_caseA4.jl which is also up to date now.
I propose we use this as a test for our further developments, say,
the objective function value has to be around 4.995e9

`run_spine_model` now also collects outputs for each window into
a dictionary that is then written to the output url using
`write_results`.