Outputs saved only until `window_end` to avoid output writing confusion.
This solution is not ideal for debugging, since it is occasionally useful for seeing the full extent of the solution, not just the part that fits into the `rolling_window`. However, this was by far the easiest way I could think of to avoid time step ambiguity when writing the time series into the database.