Simplify the computation of history time slices
At the beginning of each window, when generating the temporal
structure but *before* overwritting the current_time_slice set,
we generate the history. At this point current_time_slice is still
the one from the previous window so it's good for calculating the
history. With this, we don't need to initialize and maintain a
dedicate TimeSliceSet just for the history.

I did some tests and this still works ok without damaging performance.