Let `to_time_slice` also return history time slices
We have a convenience functor called `to_time_slice` that is used
to translate any time slice *which may not be in the model*,
into a time slice *in* the model. Dynamic constraints that look
several time steps in the past use this functor. I think now these
are min_down_time, min_up_time, and fix_ratio_out_in_trans.

Until now, this functor wasn't returning 'history' time slices,
because we just forgot to include these when generating the functor.
This commit fixes this issue.