Trick to avoid KeyError when accessing variables
As it turns out, it is much cheaper to handle the KeyError on
the `getindex` function than to go a second time through the
`variable_indices` function to validate indices.

The approach I'm using here consists on returning zero whenever
a variable is accessed with an invalid index tuple. Is it safe?