Correct way to avoid KeyError when accessing variables
Historically, we've been using `{variable}_indices` functions
to get valid indices to access our variable dictionaries. This
guarantees we never generate `KeyError`s but it's slower than
the dirty-trick in commit a85dc164b5598c03d7c4095486db57117dbbca5f

But now we're using the more correct approach that relies on
`{variable}_indices`. This seems more sane as we're in a
development phase where we always want to spot errors.