Introduce 'prehistory' time slices
Currently, if the user wants to fix variable values for time slices
*before* the model start, they need to adjust the model start and
create a dedicated temporal block for this pre-history time period.

The current commit introduces a function called `init_time_slice`
that can be used to automatically generate time slices for that
period. The idea is to start using this function in the future
to improve things.