Associate prehistory time slices to all temporal blocks
Previously we attached prehistory time slices to a block called
"prehistory", but this doesn't make much sense and doesn't work so
well. Now we just attach them to all blocks in the model so we can
blend them more easily in the forthcoming rolling structure.

We also fix some minor problems with the construction of
prehistory time slices.