Start reusing variables
We need to check how much better is performance when reusing
the same JuMP model instead of creating a new one for every window.
This commit is a first step towards that.

Basically we use the new API from SpineInterface to `roll`
TimeSlices. Rolling the model forward is achieved by calling `roll!`
on the same TimeSlices, rather than generating new ones.
This means the `m.ext[:variables]` can stay the same.

We implement shortcut functions to create, fix, and save variables.