Improve treatment of history variables
History entries in the variable Dict need to be actual JuMP variables,
otherwise the constraint does not see changes in their values.
These variables need to be fixed to the result of the corresponding
current index, using `JuMP.fix`.

Note that calling `JuMP.fix` means that we can't call `JuMP.value`
on another value afterwards. `JuMP.fix` makes the owner model lose
it's condition of 'optimized'. So we actually need an intermediate
storage, that is, `m.ext[:results]`.