#142 `node_state_balance` not looped over `node_state_indices` twice.
For whatever reason, if I included the `node_state_indices` function also in the second for loop, the constraint was generated for a ludicrously large combination of time steps. This seems to fix the issue, but is not ideal since there is no guarantee that there is a `node_state` variable on the `t_before` time step now.