Provide missing `add_to_expression!` methods
`add_to_expression!` gets called behind the scenes by
`JuMP.@constraint` and `JuMP.@objective`, and it looks like a big
part of the api is built around it. We need to provide all
variations of `add_to_expression!` involving our custom type
(`Call`) to make sure our constraint code runs.
This might be tricky as I already missed a couple, but it should
be possible to get it right with some more work.
Not sure if this commit ends it though...