Introduce generate_special_handlers
This function is called at the beginning of `run_spine_model`
and takes care of everything related to 'expand' the db contents.
It injects the `direction` object class, together with
`unit__node__direction` and `connection__node__direction`
and moves parameters around so it all works transparently.

This function also replaces the old `generate_variable_indices`
since it falls within the same type of functionality -> creating
stuff that wasn't in the db for convenience.