Get rid of direction in a different way
Turns out there's different options for doing this. Commit c396aa92
explicitly introduced connection__node__direction and
unit__node__direction, whereas here we just extend `unit__from_node`
and `connection__from_node` with the appropriate `direction` object.
This needed to change `SpineInterface.RelationshipClass` so all the
fields can be modified, but it's not a big deal. The code now looks
simpler and that's a good sign.