#141 Basic stochastic constraints for testing, model runs. WIP.
The model now generates a basic set of stochastic constraints using the stochastic structure. This version should in principle be able to generate deterministic models using a deterministic `stochastic_structure`, but I haven't tested it as such yet. More complicated `stochastic_structures` are likely to result in infeasibilities.

Also added a lot of `TODO` notes into the code to remind myself about all the aspects that still need to be implemented, like stochastic parameters.