Introduce variable `node_injection`
The purpose here is to reuse expressions corresponding to the net
node injections. The same expressions were built several times
in our ptdf constraint, basically once per connection, leading to
poor performance.

Using a variable that is equal to each of these expressions
guarantees that we build and update them only once, or as fewest
times as possible. I don't know if something like this complicates
things for the solver, though, but it greatly improves model
building time.