Introduce `expr_sum`
This is reaction to a JuMP warning we were getting because of
using `reduce`. The warning says something like 'the addition
operator has been used several times in JumP expressions and this
is suboptimal`.

So `expr_sum` now adds element in an iterable *in-place*, as
recommended by JuMP. At a first impression, it doesn't seem much
faster than `reduce`, but I guess it's ok to try and do things
by the books?