using t_overlaps instead of t_long for units_on
If the granularity of the flow variable is lower than the units_on's, units_on are summed over inherent time_slices. 
Vise versa, the "closest" units_on is used, if flow is of higher granularity (in this case, units_on is weighed with the flows duration! -> see min(duration(t1),duration(t)).