Introduce StochasticPathFinder type
Instead of creating `all_stochastic_paths` as a global and then
passing it back to the `active_stochastic_paths` function,
we introduce a new type, `StochasticPathFinder`.

This type has a unique field called `all_stochastic_paths`
that holds the paths, and overrides the Julia call operator to
perform the job of `active_stochastic_paths`, but receiving only
the `active_scenarios` argument.

We also prepend the underscore to 'internal' methods so it's easier
to debug this in the future.