Remove unnecessary (?) calls to `indices` in `_stochastic_DAG`
We don't expect the scenario tree to be really large, do we?
But in that case, calling `indices` for each `scenario` could
become expensive. Instead, we now use the `_strict` keyword
argument to the `Parameter` call and compare the result to `nothing`
to see if we need to spawn child scenarios.

I hope this works and there wasn't a strong reason for doing it
the other way!

NOTE: We should consider renaming `stochastic_scenario_end`
to `stochastic_scenario_duration`.