Twist a little bit the style in `stochastic_structure`
I don't know if this is really a contribution, but here it goes
anyways:

1. Favor building Dicts using the comprehension syntax rather than
the for loop. As far as I understand, this should be better for
performance since the `Dict` is resized just once as opposed to
once per element. That said, I don't think this part of the code
will ever become the bottleneck.

2. Keep removing the return keyword at the end of the function.
In the rest of the package, we more or less consistently just
use return when returning *in the middle* of the function.
This is questionable though, and I know e.g. JuMP prefers to always
use return. I don't know what's best but I like consistency.