Simplify `unit_stochastic_scenario_weight` and remove unused function
Julia allows to define a function with keyword arguments without
specifying their defaults, so we use that. Julia will complain
automatically if any of those kwargs is missing.

Also, I don't think we're calling `unit__stochastic_scenario` anywhere
in the code, so I'm getting rid of it. I hope it's OK.