Use the comprehension syntax for generating constraint indices
I've been doing experiments, and measured that building arrays
with the comprehension syntax can be six time faster than using
push! in a for loop. Julia is deep, but I imagine it has something
to do with the allocations (push! need to allocate a new array
every time or something like that).

So I just rewrote the constraint_..._indices functions with
comprehension syntax. Might be a bit more difficult to follow,
but I think it's worth it (even if for the test system there's no
measurable speed up at all...)