Try to use a different uuid for SpineOpt
The registry seems broken at the moment and it looks like
packages shouldn't be renamed like we did. Instead, we need
to generate a new package uuid and I'm trying to do that at the moment.

I'm not sure what are the full repercussions of doing this for
our users and developers. I first want to know if it helps fixing
the registry and will come back to that if it does.