Add first unit tests!
Re #79

Finally after about 1 year we have our first couple of tests.

To run them, just include `tests/runtests.jl` in your julia session.

The next step is to tell travis to run tests automatically and
email the committer whenever tests fail.
Until that happens, it would be a good idea to run tests
before doing something critical to dev
to make sure we're not breaking the basics.

And of course, we need to write much more tests! This commit just
includes a couple of very very basic ones.

I think tests will also be very useful for us to realize how
`run_spineopt` really works - what's the minimum that needs to be
in the db for things to happen. In this sense, maybe the user guide
and the test suite could follow a similar pattern
(do *at least* A in order for B to happen).

NOTE: I modified a couple of default values in the template to help
writing tests, but I think it also makes sense in the broad sense.
I also modified the StochasticPathFinder callable so it accepts
empty arrays as input.