Generate indices for all t_befores in dynamic constraints
Also, if no `t_before`, we don't generate any indices for the
`t_after` and, subsequently, no constraint either.

With this, we also get rid of the funny looking
`_non_empty_t_before_t` and `take_one_t_before_t` in the
`constraints_common.jl` file.

We didn't realize, at the moment
of implementing the stochastic structure for the first time,
that there could be more than one t_before for the same t_after,
and that it was ok to not have any t_before for some t_after.