Get the correct variable indices for the history
Previous to having the stochastic structure in place, we were able
to compute a 'history' index corresponding to a normal (non-history)
one, simply by replacing the `t` field in the tuple.

But with the stochastics, we also need to replace the
`stochastic_scenario` field, because history indices don't generally
have the same scenario as their non-history counterpart
(they always have all the root scenarios, in fact).