Fix gaps treatment in `to_time_slice`
From Dev; 
Context: We need to fill gaps in between temporal blocks so `to_time_slice` always returns something. We do this by building *bridge* blocks that span the gaps, and map everything to the last time slice in the preceding block (questionable, but it's what we have for now).  The problem is we were bridging not only over gaps, but also over temporal blocks themselves, leading to wrong results. As `to_time_slice` is used to generate `t_in_t`, `t_overlaps_t`, and `t_before_t`, those were also damaged.  In this commit we realize that we actually need one TimeSliceMap for each gap we want to fill. Hopefully it works better. It has been tested with a simple temporal structure consistent of two blocks, one with double the resolution than the other.  More tests will be added soon.