Adding `check_node_object()` and `check_node__temporal_block()` functions.
Now, the user should be able to start with an empty SpineOpt database (with the template), and get understandable error messages until `run_spineopt()` completes without issues. Note that the resulting system will not make a feasible model for the solver, but at least the user doesn't run into Julia method errors before seeing what the function is attempting to do.