Advance support for having one temporal structure per model object
IMPORTANT: From this commit on, `model__temporal_block` relationships
need to be defined in the db for things to work (otherwise the model
won't see any temporal structure).

Now all test pass with the new setup (one temp struct per model).
Of course, and perhaps a bit unfortunately,
all functions that query the temporal structure now need the model
as an additional argument. This includes the functions that return
variable indices as they query the temporal structure under the hood.

We also reorganize code in run_spineopt so the different operations
are more compartmentalized and we can see the high level picture
by looking at the run_spineopt function.