Try to fix style in `constraint_res_minimum_node_state.jl`
There were problems with the indentation, spaces, and maximum line
length that now are hopefully fixed.

Also, the parameter `res_start_up_cost` is now added to the template
so things work.

In addition I tried to clean up a division of time periods in the
body of the constraint, basically I'm trying to use Julia's standard
period operations instead of going through our TimeSlice api.
I believe it might be simpler now.