Added a new `optimize` keyword to `run_spineopt()` to control whether solver is run for debugging.
This keyword is assumed to be `true` so that the problem is optimized by default. However, it is occasionally useful to be able to skip running the solver when debugging, either simply to save time or to bypass potential solver issues. For now, the current implementation doesn't roll the problem if `optimize=false`, something that might need to be changed later on.