Exclude the concerned node from the internal nodes in node balance
Re #202

This is basically to enable sink connections that just take
energy from a node. If the node is considered as one of its own
internal nodes (which is questionable per se) the connection
automatically becomes 'internal'.

I think it's good to keep supporting this option of using sink
connections.