#210 Allow filtering the `<constraint>_indices` functions with (most) of their indices using keyword arguments.
Essentially, this allows similar filtering of the `<constraint>_indices` functions than what we have in the `<variable>_indices` functions. I've also rearranged the `<constraint>.jl` files so that the `add_constraint!()` function is always at the very beginning of the file, while the indexing functions are after it. There are still a couple of filtering options that are not possible due to how the indexing functions have been written, which I've tagged using `#210` comments.