Pay attention to the commodity when doing group nodal balance
Re #202

In group nodal balance, we want to ignore connection flows that are
'internal' to the group *for a given commodity*.

So if a connection has flows to a node that is not part of the group
but is from a *different* commodity, we still ignore it.

I think this is the most explicit way to achieve the desired
functionality.