Fix the tests for minimum up and down time with non spinning reserves
I'm mot entirely satisfied, first and foremost because the
test doesn't work for some 'complex' values of the parameter
(e.g. 45 minutes, 75 minutes).

And secondly, because of the way the indices of
nonspin_units_shut_down are generated, one needs to define
the max_res_shutdown_ramp (or max_res_startup_ramp parameter)
with a specific node. Namely, the node has to have the
same temporal and stochastic structure as the unit,
otherwise there's no indices being generated....