Fix indexing of unit constraint
For some reason we were ignoring one flow direction (to_node).
Also, we weren't handling properly the case where there's no flow
variable for some of the units. To expand, sometimes the user wants
to create a unit-constraint only using `units_on` variables (not
`unit_flow`) and this was not possible because of how the indexes
were determined.