Add basic_spineopt_template.json
The idea of this template is to provide the basic structure so users
can start directly creating units, nodes, and connections.
The model, temporal_block, stochastic_structure, stochastic_scenario and
all relationships therein are already provided.

Let's see if this is useful and if we want to keep it. My idea is to
test it with KTH students in a workshop that will be help in a couple
of weeks.