Add support for migration and rename unit_constraint -> user_constraint
Re Spine-project/SpineOpt.jl#120

We now support automatic migration using a simple versioning system.
The version is stored in the *default_value* of the `version`
parameter under the `settings` object class. This way we don't need
to mess with multiple `settings` objects and parameter values.

The first migration is also implemented to rename unit_constraint
to user_constraint. This happens automatically when one calls
`run_spineopt(url; upgrade=true)`. After this, the version will be 2
(we assume that 1 was the historic, unversioned version).

So please don't use the python script to update your dbs,
and don't do it manually, otherwise the automatic version detection
and migration will be broken.