Fix tests
Changes to SpineInterface (specifically, to the order in which
relationships from the db are collected, now alphabetically,
previously by insertion time) have exposed our weakness (?) at
determining sending nodes in the context of ptdf analysis.
Basically, we take the second node in the connection__from_node
result as the sending node. This might be ok thought...