Speed up saving model results
We do two things:
1. Don't regenerate indices to save variable values, they are already
available from the dictionary holding the variable
2. Don't regenerate objective terms to save their values. Instead,
save the original expressions in m.ext[:spineopt] and read them from
there to compute their values. This needs to split the expressions
into two parts: in-window, and beyond-window. The in-window part is
used to save results, but the sum of the two is used in the objective
function. This is to keep the previous functionality intact.