Introduce internal_fix_value to manage benders fixing indep. of users'
It's not a very good idea to use the DB fix_value parameter to fix
master problem solution in the subproblems. This is because we do not
know which type of parameter the user will provide with fix_value,
and it is very difficult to handle all possible cases in a merge.

For this reason, we use the internal `things_invested_available_bi`
parameter to do the fixing (which we can modify at will). Works better.