Fix duals calculation for Benders
Previously, we were computing the marginal value of the units
investment as the reduced cost of the units_on variable.
This is a bit non-intuitive - it should be the reduced cost of
units_invested_available.

But units_invested_available didn't work for us reliably,
and I think I just found out why:
when computing the duals, if units_on
was discrete in the original problem, we would fix its value
in the relaxed problem. That means more investments wouldn't be
usable (because the units_on was already fixed) resulting in reduced
cost of units_invested_available being zero.

This commit rationalizes the whole thing. The marginal value of
the unit investments is obtained from the reduced cost of
units_invested_available, as it should be. To avoid it being zero,
we don't fix discrete variables to calculate duals - we just relax them.

Note than when using CPLEX, we can't use the internal routine to solve
the relaxed problem to calculate duals, because it fixes the discrete
variables and we just can't control it. CPLEX wasn't working properly
for Benders anyway - with this commit, it works.