Improve consistency between marginal values and subproblem obj. value
Since marginal values are used to compute the improvement to the
subproblem objective value that is possible from investments,
both marginal values and sp obj value need to be computed following
the exact same rule.
The rule itself is not as relevant as the consistency.
The current calculation is,

    (value excluding look-ahead) * window_weight,

for windows 1 to second last, and,

    (value excluding look-ahead) * window_weight + (value in look-ahead)

for the very last window.

The reason to include the look-ahead in the last window is so that
the benders algorithm yields the same solution as the non-benders
(which includes the look-ahead in the objective function).

The reason to *exclude* the look-ahead in all previous windows is to
avoid any double counting.