Fix documentation
In my opinion there were slight inaccuracies in the section about
Parameter type, that I decided to try and amend. I am not sure I
succeeded, and probably we'd need to revise this and eventually
expand it to something really comprehensive that properly covers the
self-updating optimisation models; but for now I think it is ok.