Rename drag_and_drop to expand_empty_chapters!
Maybe it's only me, but I couldn't understand what we meant by
drag-and-drop until I fully read and understood the function.
I decided, rather arbitrarily, to rename it to
expand_empty_chapters! which to me seemed more direct.
Feel free to rename it back to drag_and_drop if you like it more!

I also modified the function slightly to act in place and use
some built-in path manipulation functions which seemed useful.
One, maybe substantial difference is that instead of nothing,
we need to specify an empty list to activate the automatic expansion.