Introduce sp_windows to advance work in mp-min-res-gen constraint
Also, fix double counting in the uaf term: the uaf for each window
was multiplying the potential investments increase for the whole model.
But the uaf for each window should only multiply the potential
investments increase for that window.

If the investments resolution is larger than the window then it doesn't
matter I guess, but anyways, this seems more correct.