Improve treatment of representative time slices
Previously, if the same t was represented by more than one
representative ts, only the latter would stay, leading to bugs.
Now we also help the user in case they created a mapping for a
temporal block but forgot to associate the
representative temporal block with their node or unit.