Rewrite the time slice section of the docs
Revert if needed, but I found that this section was a bit too honest.
I rewrote it in a more formal way (or at least tried to).
My preference is that user-facing documentation looks as
'professional' as possible.

But maybe I made it less clear, I don't know.
The previous text had a lot of information that I believe belongs
more in a Julia tutorial, but perhaps it is good to have it here
anyways?

Anyways, I'm open to changing this as needed.