Fix units_started_up coefficient in unit flow ratio
The unit_start_flow was being multiplied by the duration of the time
slice, as if this flow was happening at every hour of the day for
instance. But the correct interpretation should be that it flows only
once regardless of the resolution at which the unit_flow variable
is being tracked.

This also keeps consistency with the unit_pw_heat_rate constraint
so the behaviour is unaltered compared to old models that used
that constraint.