Don't create the model in prepare_spineopt
This is to support distributed execution of several spineopt instances.

Long explanation: so at the moment it is not possible to read the data
from the DB *and* run SpineOpt in parallel. It is only possible to do
the second part in parallel, running SpineOpt. Reading the DB in
parallel is simply not possible because if there is a stage,
then we need to apply the stage scenario filter to read the data
for the stage, and if several workers are trying to do this in parallel,
they will find the DB in an uncertain stage. For example, one worker
could set the filter and before resetting it, another worker could
come and try to set the filter again, and so on.

So the DB needs to be read in sequence, and then when all the workers
have read the DB in sequence, they can run SpineOpt in parallel.
That's the way it its.

In practice, this is now possible by calling
`prepare_spineopt(url_in)` in sequence,
and then calling `m = create_model()`
followed by `run_spineopt!(m, url_out)` in parallel.

But before this commit, the model was created and returned by
`prepare_spineopt` which was not good, because then one would need
to fetch the model from the worker and then send it back,
and the model object in JuMP doesn't support this.