Set add_bridges=true by default
This is mainly because some solvers like the CPLEX solver don't support
constraints of the form a <= expression <= b natively, so the bridges
are necessary to make it work.
In our model the LODF constraint is of this form.
Also, add_bridges=true is the default for the Model constructor in
JuMP, so it's more consistent to use the same default.

I believe add_bridges=false was aimed at improving performance but
we cannot do it at the cost of breaking some previous workflows that
were working (including the LODF constraint with CPLEX).
Perhaps with more time in the future we can do a full assessment of
which constraints require bridges in our model and rewrite them.
This is under the assumption that using JuMP bridges is more
expensive than rewritting the offending constraints ourselves,
which is a question for me.